moving all test files into a separated folder like requested in issue #9
